refactor julia plugin & add tests for 0801 & 0804
